Refactor layouts into lib/components/layouts
