--- src/lib/components/stories/Logo.stories.svelte
@@ -1,14 +1,0 @@
-<script context="module" lang="ts">
-  import { Story } from "@storybook/addon-svelte-csf";
-  import Logo from "../common/Logo.svelte";
-
-  export const meta = {
-    title: "Components / Common / Logo",
-    component: Logo,
-    parameters: { layout: "centered" },
-  };
-</script>
-
-<Story name="Logo" id="logo">
-  <Logo />
-</Story>
